refactor(tests): merge panel fixtures and drop dead harness layers
- bookmark/persons/shortcuts fixtures differed only by panel name, and each
  aliased its page onto a second fixture that just forwarded it. One
  panel-fixture exports the three tests directly.
- test-utils re-exported eight shared helpers "for convenience"; the three
  files that used them now import from @bypass/shared/tests.
- verifyModalVisible/verifyModalClosed each had a single caller that always
  passed a test id, so the no-id branch was unreachable. Inlined into the web
  persons POM.
- The @bypass/web Playwright project collected zero tests: its testIgnore
  excluded everything under its own testDir.
- warnOnClientComponentExclusions re-read a hand-maintained exclusion list to
  warn about it.
