+++ playwright.config.ts
@@ -36,14 +36,6 @@ const config = defineConfig({
     headless: true,
   },
   projects: [
-    {
-      name: '@bypass/web',
-      testDir: './apps/web/tests',
-      testIgnore: ['**/auth.setup.ts', 'specs/**/*'],
-      use: {
-        baseURL: ciBaseUrl ?? 'http://localhost:3000',
-      },
-    },
     /**
      * Web App Setup: Runs once per test run to authenticate and cache storage.
      * This avoids repeating the login flow and data sync for every test.
